requirements.txt
flask==2.0.1
torch==1.12.1
diffusers==0.11.1
transformers==4.21.3
accelerate==0.15.0
Pillow==9.2.0